uh
so
i don't know what i did, but now NONE of my react pages work. not my original project, not this new project, not my servers, NOTHING. i get this error on all of them: "The character encoding of the plain text document was not declared. The document will render with garbled text in some browser configurations if the document contains characters from outside the US-ASCII range. The character encoding of the file needs to be declared in the transfer protocol or file needs to use a byte order mark as an encoding signature."
I don't know what i did, but i broke everything.